Skip fields with no subfields (or only empty subfields)
This is a follow on to PR #1 which fixed the issue of self-closing
subfield tags by interpreting them as an empty string. After
discussion it was decided that these are better modeled as
containing `None` and being filtered out from future use, as they
are functionally the same as a non-existant tag.

This handles them as `None` values, checking for `None` in subfields
and filtering out any subfields that meet the criteria and then
subsequently filtering out any fields that contain no subfields.

The test case has been retained from PR #1 but now expects the
request to evaluate to an empty list.